machine tester

+++ PoC/Models/ItemToDispense.cs
@@ -5,7 +5,7 @@ namespace PoC.Models
     public class ItemToDispense
     {
         public PoGProduct Product { get; set; }
-        public uint Qty { get; set; }
+        public ushort Qty { get; set; }
 
         public override string ToString() => $"{Product?.ProductUid}x{Qty}";
     }
